+++ colgm-bootstrap/src/misc.cpp
@@ -223,7 +223,7 @@ f64 str_to_num(const char* str) {
     return negative? -res:res;
 }
 
-i32 utf8_hdchk(const char head) {
+usize utf8_hdchk(const char head) {
     // RFC-2279 but now we use RFC-3629 so nbytes is less than 4
     const auto c = static_cast<u8>(head);
     if ((c>>5)==0x06) { // 110x xxxx (10xx xxxx)^1
